Move in_workspace_with_projects_file to separate util module
To silence false-positive warnings about dead code for tests that don't
use those functions.
